refactor: move agent_management to the new architecture
